+++ hutool-crypto/src/test/java/cn/hutool/crypto/test/SmTest.java
@@ -75,6 +75,20 @@ public class SmTest {
 	}
 
 	@Test
+	public void sm4TestWithCustomKeyTest2() {
+		String content = "test中文frfewrewrwerwer---------------------------------------------------";
+
+		byte[] key = KeyUtil.generateKey(SM4.ALGORITHM_NAME, 128).getEncoded();
+
+		SM4 sm4 = SmUtil.sm4(key);
+		Assert.assertEquals("SM4", sm4.getCipher().getAlgorithm());
+
+		String encryptHex = sm4.encryptHex(content);
+		String decryptStr = sm4.decryptStr(encryptHex, CharsetUtil.CHARSET_UTF_8);
+		Assert.assertEquals(content, decryptStr);
+	}
+
+	@Test
 	public void hmacSm3Test() {
 		String content = "test中文";
 		HMac hMac = SmUtil.hmacSm3("password".getBytes());