+++ hutool-db/src/main/java/cn/hutool/db/meta/Column.java
@@ -1,6 +1,7 @@
 package cn.hutool.db.meta;
 
 import cn.hutool.core.util.BooleanUtil;
+import cn.hutool.core.util.ReUtil;
 import cn.hutool.db.DbRuntimeException;
 
 import java.io.Serializable;
@@ -111,7 +112,12 @@ public class Column implements Serializable, Cloneable {
 		this.isPk = table.isPk(this.name);
 
 		this.type = columnMetaRs.getInt("DATA_TYPE");
-		this.typeName = columnMetaRs.getString("TYPE_NAME");
+
+		String typeName = columnMetaRs.getString("TYPE_NAME");
+		//issue#2201@Gitee
+		typeName = ReUtil.delLast("\\(\\d+\\)", typeName);
+		this.typeName = typeName;
+
 		this.size = columnMetaRs.getInt("COLUMN_SIZE");
 		this.isNullable = columnMetaRs.getBoolean("NULLABLE");
 		this.comment = columnMetaRs.getString("REMARKS");