+++ CHANGELOG.md
@@ -19,6 +19,7 @@
 * 【core   】     Base32增加pad支持（pr#2195@Github）
 * 【core   】     Dict增加setFields方法（pr#578@Gitee）
 * 【db     】     新加db.meta的索引相关接口（pr#563@Gitee）
+* 【db     】     Oracle中Column#typeName后的长度去掉（pr#563@Gitee）
 * 
 ### 🐞Bug修复
 * 【core   】     修复ObjectUtil.hasNull传入null返回true的问题（pr#555@Gitee）